Add catalog model APIs
