+++ pipes/catalogmodels/schemas.py
@@ -1,0 +1,113 @@
+from __future__ import annotations
+
+from datetime import datetime
+
+import pymongo
+from pymongo import IndexModel
+from beanie import Document, PydanticObjectId
+from pydantic import BaseModel, Field, field_validator
+
+from pipes.users.schemas import UserRead
+
+
+class CatalogModelCreate(BaseModel):
+    """Model schema for catalog"""
+
+    name: str = Field(
+        title="model_catalog",
+        min_length=1,
+        description="the model name",
+    )
+    display_name: str | None = Field(
+        title="display_name",
+        default=None,
+        description="Display name for this model vertex.",
+    )
+    type: str = Field(
+        title="type",
+        description="Type of model to use in graphic headers (e.g, 'Capacity Expansion')",
+    )
+    description: list[str] = Field(
+        title="description",
+        description="Description of the model",
+    )
+    assumptions: list[str] = Field(
+        title="assumptions",
+        description="List of model assumptions",
+        default=[],
+    )
+    requirements: dict = Field(
+        title="requirements",
+        default={},
+        description="Model specific requirements (if different from Project and Project-Run)",
+    )
+    expected_scenarios: list[str] = Field(
+        title="expected_scenarios",
+        description="List of expected model scenarios",
+        default=[],
+    )
+    other: dict = Field(
+        title="other",
+        default={},
+        description="other metadata info about the model in dictionary",
+    )
+    access_group: PydanticObjectId | None = Field(
+        title="access_group",
+        default=None,
+        description="Group that has access to this model",
+    )
+
+    @field_validator("description", mode="before")
+    @classmethod
+    def validate_description(cls, value):
+        if isinstance(value, str):
+            return [value]
+        return value
+
+
+class CatalogModelUpdate(CatalogModelCreate):
+    pass
+
+
+class CatalogModelRead(CatalogModelCreate):
+
+    created_at: datetime = Field(
+        title="created_at",
+        description="catalog model creation time",
+    )
+
+    created_by: UserRead = Field(
+        title="created_by",
+        description="user who created the model in catalog",
+    )
+
+
+class CatalogModelDocument(CatalogModelCreate, Document):
+    created_at: datetime = Field(
+        title="created_at",
+        description="catalog model creation time",
+    )
+    created_by: PydanticObjectId = Field(
+        title="created_by",
+        description="user who created the model in catalog",
+    )
+    last_modified: datetime = Field(
+        title="last_modified",
+        default=datetime.now(),
+        description="last modification datetime",
+    )
+    modified_by: PydanticObjectId = Field(
+        title="modified_by",
+        description="user who modified the project",
+    )
+
+    class Settings:
+        name = "catalogmodels"
+        indexes = [
+            IndexModel(
+                [
+                    ("name", pymongo.ASCENDING),
+                ],
+                unique=True,
+            ),
+        ]
