+++ pipes/models/schemas.py
@@ -45,59 +45,6 @@ class ScenarioMapping(BaseModel):
         return value
 
 
-class CatalogModelCreate(BaseModel):
-    """Model schema for catalog"""
-
-    name: str = Field(
-        title="model_catalog",
-        min_length=1,
-        description="the model name",
-    )
-    display_name: str | None = Field(
-        title="display_name",
-        default=None,
-        description="Display name for this model vertex.",
-    )
-    type: str = Field(
-        title="type",
-        description="Type of model to use in graphic headers (e.g, 'Capacity Expansion')",
-    )
-    description: list[str] = Field(
-        title="description",
-        description="Description of the model",
-    )
-    assumptions: list[str] = Field(
-        title="assumptions",
-        description="List of model assumptions",
-        default=[],
-    )
-    requirements: dict = Field(
-        title="requirements",
-        default={},
-        description="Model specific requirements (if different from Project and Project-Run)",
-    )
-    expected_scenarios: list[str] = Field(
-        title="expected_scenarios",
-        description="List of expected model scenarios",
-        default=[],
-    )
-    other: dict = Field(
-        title="other",
-        default={},
-        description="other metadata info about the model in dictionary",
-    )
-
-    @field_validator("description", mode="before")
-    @classmethod
-    def validate_description(cls, value):
-        if isinstance(value, str):
-            return [value]
-        return value
-
-    class Settings:
-        name = "model_catalog"
-
-
 class ModelCreate(BaseModel):
     """Model schema"""
 
@@ -198,37 +145,6 @@ class ModelRead(ModelCreate):
     )
 
 
-class CatalogModelDocument(CatalogModelCreate, Document):
-    created_at: datetime = Field(
-        title="created_at",
-        description="project creation time",
-    )
-    created_by: PydanticObjectId = Field(
-        title="created_by",
-        description="user who created the project",
-    )
-    last_modified: datetime = Field(
-        title="last_modified",
-        default=datetime.now(),
-        description="last modification datetime",
-    )
-    modified_by: PydanticObjectId = Field(
-        title="modified_by",
-        description="user who modified the project",
-    )
-
-    class Settings:
-        name = "catalogmodels"
-        indexes = [
-            IndexModel(
-                [
-                    ("name", pymongo.ASCENDING),
-                ],
-                unique=True,
-            ),
-        ]
-
-
 class ModelDocument(ModelRead, Document):
     context: ProjectRunObjectContext = Field(
         title="context",
@@ -242,20 +158,20 @@ class ModelDocument(ModelRead, Document):
     # document information
     created_at: datetime = Field(
         title="created_at",
-        description="project creation time",
+        description="model creation time",
     )
     created_by: PydanticObjectId = Field(
         title="created_by",
-        description="user who created the project",
+        description="user who created the model",
     )
     last_modified: datetime = Field(
         title="last_modified",
         default=datetime.now(),
-        description="last modification datetime",
+        description="model's last modification datetime",
     )
     modified_by: PydanticObjectId = Field(
         title="modified_by",
-        description="user who modified the project",
+        description="user who modified the model",
     )
 
     class Settings:
